RTE Model refactoring: (#494) (#776)
* RTE Model refactoring:

derive RteItem from XmlTreeItem<RteItem>

make VISIT result class enum

remove duplicated methods from RteItem

use size_t instead of int for collection sizes

add RteComponentInstance::GetComponent() override to simplify logic

simplify XmlTreeItem::AddChild()

+++ libs/rtemodel/include/RteComponent.h
@@ -51,7 +51,7 @@ public:
    * @brief get number of files described in this component
    * @return number of <file> elements
   */
-  int GetFileCount() const;
+  size_t GetFileCount() const;
   /**
    * @brief get <files> element
    * @return pointer to RteFileContainer representing container for RteFile items
